requirements.txt
gradio~=5.23.1
langchain-groq~=0.3.2
langchain-openai~=0.3.11
python-dotenv~=1.1.0